requirements.txt
docarray==0.40.0
ipython==8.0.1
langchain==0.1.19
langchain-community==0.2.0
langchain-openai==0.1.6
openai==1.27
python-dotenv==1.0.1
tiktoken==0.6.0
yt_dlp==2024.8.6